fix(db): widen volume/shares to BIGINT — stops scan-write overflow + Sentry flood (#300)
* fix(db): widen count columns to BIGINT to stop 32-bit overflow

The scan-tick bulk write into `tickers` was failing every cycle with
psycopg NumericValueOutOfRange: a high-turnover name (ADTX ~5.28B shares)
exceeds the 32-bit INTEGER max, and one overflowing row aborted the entire
batch — so scores stopped persisting AND Sentry got flooded with thousands
of identical DataErrors (exhausting the 5k/period error quota).

- tickers.volume INTEGER -> BIGINT (the active fire)
- institutional_holdings.shares -> BIGINT (a fund can hold >2.1B shares)
- ipo_events.shares_offered -> BIGINT (latent same-class overflow)
- migration 0034 (Postgres-only; SQLite INTEGER is already 64-bit)
- regression test asserting all three stay BigInteger


* style(models): drop now-unused Integer import in ticker.py (ruff F401)


---------

+++ backend/app/models/holdings.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 
 from datetime import datetime
 
-from sqlalchemy import DateTime, Float, Integer, String, func
+from sqlalchemy import BigInteger, DateTime, Float, Integer, String, func
 from sqlalchemy.orm import Mapped, mapped_column
 
 from app.db import Base
@@ -25,7 +25,8 @@ class InstitutionalHolding(Base):
     cik: Mapped[str] = mapped_column(String(20), nullable=False, index=True)
     symbol: Mapped[str] = mapped_column(String(20), nullable=False, index=True)
     value_usd: Mapped[float] = mapped_column(Float, nullable=False, default=0.0)
-    shares: Mapped[int] = mapped_column(Integer, nullable=False, default=0)
+    # BigInteger: large institutions can hold >2.147B shares of a single name.
+    shares: Mapped[int] = mapped_column(BigInteger, nullable=False, default=0)
     percent_portfolio: Mapped[float] = mapped_column(Float, nullable=False, default=0.0)
     fetched_at: Mapped[datetime] = mapped_column(
         DateTime(timezone=True), server_default=func.now(), nullable=False,
